backend/requirements.txt
flask
flask-sock
flask-cors
flask-sqlalchemy
oauthlib
authlib
requests
httpx
mcp
google-genai
google-api-python-client
google-auth
python-dotenv
